Refactored metrics widget
