+++ frontend/src/components/__tests__/MetricsWidget.test.tsx
@@ -1,31 +1,24 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { WidgetType, MetricsWidget } from "../../models";
-import MetricsWidgetComponent from "../MetricsWidget";
+import { MemoryRouter } from "react-router-dom";
+import MetricsWidget from "../MetricsWidget";
 
-jest.mock("../../hooks");
-window.URL.createObjectURL = jest.fn();
+test("renders the title and summary of the metrics preview component", async () => {
+  const { getByText } = render(
+    <MetricsWidget title="test title" metrics={[]} metricPerRow={3} />,
+    { wrapper: MemoryRouter }
+  );
+  expect(getByText("test title")).toBeInTheDocument();
+});
 
-const metrics: MetricsWidget = {
-  id: "123",
-  name: "Test metrics",
-  dashboardId: "abc",
-  updatedAt: new Date(),
-  widgetType: WidgetType.Metrics,
-  order: 0,
-  showTitle: true,
-  content: {
-    title: "Test metrics",
-    s3Key: {
-      raw: "123.json",
-      json: "123.json",
-    },
-    datasetId: "123",
-    oneMetricPerRow: false,
-  },
-};
-
-test("renders an metrics with title", async () => {
-  const { getByText } = render(<MetricsWidgetComponent widget={metrics} />);
-  expect(getByText("Test metrics")).toBeInTheDocument();
+test("metrics preview should match snapshot", async () => {
+  const wrapper = render(
+    <MetricsWidget
+      title="test title"
+      metrics={[{ title: "test title", value: 1 }]}
+      metricPerRow={3}
+    />,
+    { wrapper: MemoryRouter }
+  );
+  expect(wrapper.container).toMatchSnapshot();
 });
