+++ frontend/src/containers/AddMetrics.tsx
@@ -11,7 +11,7 @@ import {
 import { useDashboard } from "../hooks";
 import BackendService from "../services/BackendService";
 import Breadcrumbs from "../components/Breadcrumbs";
-import MetricsPreview from "../components/MetricsPreview";
+import MetricsWidget from "../components/MetricsWidget";
 import TextField from "../components/TextField";
 import Button from "../components/Button";
 import MetricsList from "../components/MetricsList";
@@ -267,7 +267,7 @@ function AddMetrics() {
             </div>
             <div className="grid-col-6">
               <h4 className="margin-top-4">Preview</h4>
-              <MetricsPreview
+              <MetricsWidget
                 title={showTitle ? title : ""}
                 metrics={metrics}
                 metricPerRow={oneMetricPerRow ? 1 : 3}
